Specify the timeout as a Double, allowing to express sub-second timeouts

+++ src/HipSpec/Params.hs
@@ -22,7 +22,7 @@ data Params = Params
 
     , processes           :: Int
     , batchsize           :: Int
-    , timeout             :: Int
+    , timeout             :: Double
     , provers             :: String
     , methods             :: String
     , readable_tptp       :: Bool
